+++ src/dev/dir-structure/examples/writing_dir_children.rs
@@ -16,19 +16,22 @@ pub fn main() -> Result<(), Box<dyn std::error::Error>> {
     let path = example_dirs::get_example_dir_path("temp_children");
 
     let dir = Dir {
-        children: DirChildren::with_children_from_iter(
-            path.clone(),
-            [
-                DirChild::new("ab.txt", "Hello".to_owned()),
-                DirChild::new("cd.txt", "world!".to_owned()),
-            ],
-        ),
+        children: [
+            DirChild::new("ab.txt", "Hello".to_owned()),
+            DirChild::new("cd.txt", "world!".to_owned()),
+        ]
+        .into_iter()
+        .collect(),
     };
 
+    println!("Writing children...");
     dir.write(&path)?;
+    println!("Done writing.");
 
+    println!("Verifying...");
     assert_eq!(fs::read_to_string(path.join("ab.txt"))?, "Hello");
     assert_eq!(fs::read_to_string(path.join("cd.txt"))?, "world!");
+    println!("Verified.");
 
     Ok(())
 }